+++ app/Http/Controllers/API/ProfileController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\API\BaseController as BaseController;
 use App\Profile;
 use App;
 use Validator;
+use Illuminate\Support\Facades\DB;
 use App\Http\Resources\Profile as ProfileResource;
 
 class ProfileController extends BaseController
@@ -177,6 +178,14 @@ class ProfileController extends BaseController
 
     public function delete($id)
     {
+        $pid = DB::table('profiles')->where('id', $id)->pluck('user_id')->first();
+        $usid = auth()->user()->id;
+
+        if($pid!=$usid)
+        {
+            return $this->sendError('Sorry, delete your own profile.');
+        }
+
         //$profile->delete();
         Profile::destroy($id);
